+++ RED/Interfaces/IMetadata.cs
@@ -2,7 +2,7 @@
 {
     public interface IMetadata
     {
-        byte Id { get; set; }
+        ushort Id { get; set; }
         string Name { get; set; }
         string Description { get; set; }
         string Datatype { get; set; }
